addyy worked a lot of problems in 6hrs

+++ libft/cFiles/ft_strncmp.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_strncmp(const char *s1, const char *s2, int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	int i;
 	int diff;
